requirements.txt
# Core dependencies for MusicGen research project
# Python 3.9+ required

# Deep Learning
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0

# MusicGen and AudioCraft
audiocraft>=1.3.0
transformers>=4.35.0
accelerate>=0.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
umap-learn>=0.5.4

# Interpretability tools
einops>=0.7.0
scikit-learn>=1.3.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
jupyter>=1.0.0
ipywidgets>=8.1.0
pytest>=7.4.0

# For causal analysis
networkx>=3.1
pygraphviz>=1.11  # Optional: for graph visualization

# For downloading datasets
datasets>=2.14.0
huggingface-hub>=0.17.0
gdown>=4.7.0

# Audio classification (for evaluation)
laion-clap>=1.1.0  # For CLAP scores
